Atualização do modelo Produto (inclusão da data de fabricação e data de vencimento), implantação da função de adicionar items a receita.

+++ Gracibolos/src/main/java/br/com/gracibolos/jdbc/teste/ProdutoExcluir.java
@@ -11,7 +11,7 @@ public class ProdutoExcluir {
 		ProdutoDao dao = new ProdutoDao();
 		
 		Produto produto = new Produto();
-		produto.setId(4);
+		produto.setId((long) 4);
 				
 		try {
 			if(dao.excluir(produto) == true){
